// frontend/src/api/schema.d.ts
export interface paths {
    "/api/v1/users/me": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** 내 프로필 조회 */
        get: operations["getMyProfile"];
        /** 내 프로필 수정 */
        put: operations["updateMyProfile"];
        post?: never;
        /** 내 계정 탈퇴 */
        delete: operations["deleteMyAccount"];
        options?: never;
        head?: never;
        /** 내 프로필 수정 */
        patch: operations["updateMyProfile_1"];
        trace?: never;
    };
// ...
    "/api/v1/walks/{walkId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** 산책 기록 조회 */
        get: operations["getWalk"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
// ...
export interface operations {
// ...
    delete: {
        parameters: {
            query?: {
                confirmSelfLockout?: boolean;
            };
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}